+++ lib/saluki-io/src/net/unix/mod.rs
@@ -1,6 +1,8 @@
 use std::{io, os::unix::fs::FileTypeExt as _, path::Path};
 
+use bytes::BufMut;
 use socket2::SockRef;
+use tokio::net::{UnixDatagram, UnixStream};
 use tracing::trace;
 
 mod ancillary;
@@ -8,15 +10,14 @@ mod ancillary;
 #[cfg(target_os = "linux")]
 mod linux;
 #[cfg(target_os = "linux")]
-use self::linux::set_uds_passcred;
-
-mod recvmsg;
-pub(super) use self::recvmsg::{unix_recvmsg, unixgram_recvmsg};
+use self::linux::{set_uds_passcred, uds_recvmsg};
 
 #[cfg(not(target_os = "linux"))]
 mod non_linux;
 #[cfg(not(target_os = "linux"))]
-use self::non_linux::set_uds_passcred;
+use self::non_linux::{set_uds_passcred, uds_recvmsg};
+
+use super::addr::ConnectionAddress;
 
 /// Configures a Unix domain socket for use.
 ///
@@ -64,4 +65,29 @@ pub(super) async fn ensure_unix_socket_free<P: AsRef<Path>>(path: P) -> io::Resu
     }
 
     Ok(())
+}
+
+pub async fn unix_recvmsg<B: BufMut>(socket: &mut UnixStream, buf: &mut B) -> io::Result<(usize, ConnectionAddress)> {
+    // TODO: We technically don't need to do this for SOCK_STREAM because we can do it once when the
+    // connection is accepted, and then just do "normal" reads after that. We do still need to do it
+    // for SOCK_DGRAM, though... so we might want to actually consider updating our `socket2` PR to
+    // include support for even setting SO_PASSCRED (or the OS-specific equivalent) to also include
+    // macOS and FreeBSD (*BSD, really) so that this also works correctly for all
+    // `cfg(unix)`-capable platforms.
+
+    // We manually call `recvmsg` on our domain socket as stdlib/`mio`/`tokio` don't yet expose a way to do out-of-band
+    // reads to get ancillary data such as the socket credentials used to shuttle origin detection information.
+    socket
+        .async_io(tokio::io::Interest::READABLE, || uds_recvmsg(socket, buf))
+        .await
+}
+
+pub async fn unixgram_recvmsg<B: BufMut>(
+    socket: &mut UnixDatagram, buf: &mut B,
+) -> io::Result<(usize, ConnectionAddress)> {
+    // We manually call `recvmsg` on our domain socket as stdlib/`mio`/`tokio` don't yet expose a way to do out-of-band
+    // reads to get ancillary data such as the socket credentials used to shuttle origin detection information.
+    socket
+        .async_io(tokio::io::Interest::READABLE, || uds_recvmsg(socket, buf))
+        .await
 }
